requirements.txt
crewai
langchain-groq
pydantic
python-pptx
huggingface_hub
pillow
python-dotenv